Move order to market, remove unused code
