+++ backend/src/market/index.ts
@@ -1,24 +1,35 @@
-// import Asset from '../strategy/asset';
-import { Order } from '../simulation';
+import { uniqueId } from 'lodash';
 import Tick from '../models/tick';
+
+export class Order {
+  id: string;
+  date: Date;
+  quantity: number;
+  price: number;
+  type: string;
+
+  constructor({ date, quantity, price, type }: {
+    date: Date;
+    quantity: number;
+    price: number;
+    type: string;
+  }) {
+    this.id = uniqueId();
+    this.date = date;
+    this.quantity = quantity;
+    this.price = price;
+    this.type = type;
+  }
+}
 
 class Market {
   saveOrder?(order: Order): () => void;
-  // onAssetSell?(asset: Asset): () => void;
-  // Asset: any;
 
   constructor({ saveOrder }) {
     this.saveOrder = saveOrder;
   }
 
   buy(tick: Tick, quantity: number) {
-    // const asset = new this.Asset(tick.last, quantity, tick.timestamp);
-
-    // How to clean this up?
-    // asset.handleSell = (price) => {
-    //   this.handleAssetSell(asset, price)
-    // };
-
     return this.createOrder({
       date: tick.timestamp,
       price: tick.last,
@@ -49,18 +60,6 @@ class Market {
 
     return order;
   }
-
-  // handleAssetSell(asset: Asset, price: number) {
-  //   this.createOrder({
-  //     type: 'sell',
-  //     asset,
-  //     price,
-  //   })
-
-  //   if (this.onAssetSell) {
-  //     this.onAssetSell(asset);
-  //   }
-  // }
 }
 
 export default Market;
